Move Catch into its own internal module.
