+++ src/Control/Effect/Internal.hs
@@ -1,8 +1,7 @@
 {-# LANGUAGE DeriveFunctor, DeriveGeneric, ExistentialQuantification, KindSignatures, StandaloneDeriving, TypeOperators #-}
 module Control.Effect.Internal
 ( -- * Effects
-  Catch(..)
-, Choose(..)
+  Choose(..)
 , Empty(..)
 , Error
 , Lift(..)
@@ -13,24 +12,10 @@ module Control.Effect.Internal
 , Writer(..)
 ) where
 
+import Control.Effect.Catch.Internal
 import Control.Effect.Class
 import Control.Effect.Sum ((:+:))
 import GHC.Generics (Generic1)
-
--- | 'Catch' effects can be used alongside 'Control.Effect.Throw.Throw' to provide recoverable exceptions.
---
--- @since 1.0.0.0
-data Catch e m k
-  = forall b . Catch (m b) (e -> m b) (b -> m k)
-
-deriving instance Functor m => Functor (Catch e m)
-
-instance HFunctor (Catch e) where
-  hmap f (Catch m h k) = Catch (f m) (f . h) (f . k)
-
-instance Effect (Catch e) where
-  handle state handler (Catch m h k) = Catch (handler (m <$ state)) (handler . (<$ state) . h) (handler . fmap k)
-
 
 -- | @since 1.0.0.0
 newtype Choose m k
